Add a shebang to generate and exit immediately on failure
bash -e exits immediately if a command fails. This means it doesn't try
to run commands in an invalid state. AFAIK this is a bashism so the bash
shebang is added.

If the command runs as `sh generate.sh` it does not use the -e mode and
it'll remain POSIX compatible.

[noissue]