Adds support for running containers with a specific MCS label
If the bindings being generated need to be immediately installed inside another container,
the MCS label on the files needs to match the MCS label of that other container. This
patch enables the user to set the $PULP_MCS_LABEL environment variable (e.g. s0:c1,c2).
When this variable is present, the container for openapi-generator-cli will be started with
this MCS label.

https://opensource.com/article/18/2/selinux-labels-container-runtimes

[noissue]