Relax python-dateutil requirement
This allows python-dateutil version 2.8.1 to be installed as required by
some pulpcore dependencies. In turn this change is a shallow workaround
to prevent pip from silently breaking Pulp when installed in the same
environment (you probably should not do that, but our test suite has this as
a hard requirement).

[noissue]